JezykiSkryptowe/projekt/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Bartosz Gotowski <bartoszgotowski03@gmail.com>"]
readme = "README.md"
package-mode = false 

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.111.0"
pydantic = "^2.7.1"
SQLAlchemy = "^2.0.30"
alembic = "^1.13.1"
langchain = "^0.2.0"
langchain-community = "^0.2.0"
tavily-python = "^0.3.3"
langchain-openai = "^0.1.7"
langchain-anthropic = "^0.1.13"
python-dotenv = "^1.0.1"
sqlmodel = "^0.0.18"
pydantic-settings = "^2.2.1"
langchain-experimental = "^0.0.59"
langchainhub = "^0.1.15"
qdrant-client = "^1.9.1"
unstructured = "^0.14.2"
langchain-qdrant = "^0.1.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
